temporary removed 3d for clean-up

+++ Rndr/src/Platform/OpenGL/OpenGLVertexArray.cpp
@@ -66,7 +66,7 @@ namespace Rndr
 				ShaderDataTypeToOpenGLBaseType(element.Type),
 				element.Normalized ? GL_TRUE : GL_FALSE,
 				layout.GetStride(),
-				(const void*)element.Offset);
+				(const void*)(uint64_t)element.Offset);
 			index++;
 		}
